+++ crates/leptonica-transform/tests/bilinear_reg.rs
@@ -397,7 +397,7 @@ fn bilinear_reg_pta_basic() {
     let mut pix_mut = pix.try_into_mut().unwrap();
     for y in 0..100u32 {
         for x in 0..100u32 {
-            let _ = pix_mut.set_pixel(x, y, ((x + y) % 256) as u32);
+            let _ = pix_mut.set_pixel(x, y, (x + y) % 256);
         }
     }
     let pix: Pix = pix_mut.into();
