+++ crates/leptonica-transform/tests/projective_reg.rs
@@ -302,7 +302,7 @@ fn projective_reg_pta_basic() {
     let mut pix_mut = pix.try_into_mut().unwrap();
     for y in 0..100u32 {
         for x in 0..100u32 {
-            let _ = pix_mut.set_pixel(x, y, ((x + y) % 256) as u32);
+            let _ = pix_mut.set_pixel(x, y, (x + y) % 256);
         }
     }
     let pix: Pix = pix_mut.into();
